Rename files to use column_schema and table_schema
